Enhance list and text components with improved delete functionality and list type switching; add cursor management for better user experience

+++ src/modules/page/components/list/listComponent.tsx
@@ -7,6 +7,8 @@ import { PageContext, PageContextValues } from "../../page";
 import { useContext, useEffect, useState } from "react";
 import styles from "../../page.module.css"
 import { Mode } from "../../types/pageEnums";
+import { PageType } from "../../types/pageEnums";
+import { v4 as uuidv4 } from 'uuid';
 
 export default function ListComponent({ item }: { item: PageItem }) {
   const { pageContent, setPageContent } = useContext(PageContext) as PageContextValues;
@@ -80,7 +82,6 @@ export default function ListComponent({ item }: { item: PageItem }) {
   const handleKeyDown = (e: React.KeyboardEvent, index: number) => {
     if (e.key === "Enter") {
       e.preventDefault();
-
       if (listContent[listContent.length - 1] === ""){
         const listContentCopy = [...listContent];
         listContentCopy.pop();
@@ -96,13 +97,26 @@ export default function ListComponent({ item }: { item: PageItem }) {
       }, 1);
     } else if (e.key === "Backspace" && isCursorAtStart(e.currentTarget as HTMLElement)) {
       e.preventDefault();
+      if (listContent.length === 1 && listContent[0] === ""){
+        const pageContentCopy = [...pageContent];
+        const intemIndex = pageContentCopy.findIndex((pageItem: PageItem) => pageItem.id === item.id);
+        const previousElement = document.getElementById(pageContentCopy[intemIndex - 1].id) as HTMLElement;
+        pageContentCopy.splice(intemIndex, 1);
+        setPageContent(pageContentCopy);
+        setTimeout(() => {
+          previousElement.focus();
+          moveCursorToEnd(previousElement);
+        }, 1);
+        return
+      }
+
       deleteItem(index);
       setTimeout(() => {
         const nextElementId = `${item.id}-${index - 1}`;
         const nextElement = document.getElementById(nextElementId) as HTMLElement;
         if (nextElement) {
           nextElement.focus();
-          moveCursorToEnd(nextElement); // Mover el cursor al final del elemento
+          moveCursorToEnd(nextElement); 
         }
       }, 1);
     }
@@ -131,12 +145,21 @@ export default function ListComponent({ item }: { item: PageItem }) {
     pageContentCopy.splice(index, 1);
     setPageContent(pageContentCopy);
   }
+
+  const onChangeType = ({e, type}:{e: React.MouseEvent, type: "ordered" | "unordered"}) => {
+    e.stopPropagation();
+    const pageContentCopy = [...pageContent];
+    const index = pageContentCopy.findIndex((pageItem: PageItem) => pageItem.id === item.id);
+    pageContentCopy[index].listType = type;
+    setPageContent(pageContentCopy);
+  }
   
   const popContent = () => {
     return <div className={styles.intemPopover}>
+      <Popover content={<span style={{ color: "white" }}>Change to {item.listType === "ordered" ? "Bullet list" : "Numbered list"}</span>} color="var(--folderTextColor)">
+        <Button icon={<img src={item.listType === "ordered" ? bulletIcon : numberIcon} />} onClick={e => onChangeType({e, type: item.listType === "ordered" ? "unordered" : "ordered"})} />
+      </Popover>
       <Popover content={<span style={{ color: "white" }}>Delete this list</span>} color="var(--folderTextColor)">
-        <Button icon={<img src={bulletIcon} />} onClick={() => item.listType = "unordered"} />
-        <Button icon={<img src={numberIcon} />} onClick={() => item.listType = "ordered"} />
         <Button icon={<CiTrash />} onClick={onDelete} />
       </Popover>
     </div>
@@ -144,8 +167,8 @@ export default function ListComponent({ item }: { item: PageItem }) {
   }
 
   return (
-    <Popover  placement="topLeft" content={popContent()} color="var(--pageBarColor)" >
-      <ul id={item.id} style={{ listStyleType: item.listType === "ordered" ? "decimal" : "disc" }}>
+    <Popover  placement="topLeft" content={popContent()} color="var(--pageBarColor)"  >
+      <ul id={item.id} style={{ listStyleType: item.listType === "ordered" ? "decimal" : "disc", marginLeft: "20px"}}>
         {listContent.map((content, index) => (
           <li
             key={`${item.id}-${index}`}
